Move extensions class to Internal/Extensions.
